Implemented a basic non-blocking input/output routing with poll(2)
If there's a straggler output that locks down an input's buffer, then
the entire data flow may stall.  This is because the buffers are tied to
inputs and they get locked while the contents is being written to an
output.

TODO Decouple the buffers from inputs, maintain |inputs| + |outputs|
buffers and route by swapping to ensure everyone can stay busy on I/O.