Tries to mitigate the rebooting issue on OS X 10.11
- Throttles only if the poll result indicates no I/O is possible.
- Changes to THROTTLE_SLEEP_USEC for micro-second resolution, and uses
  100us = 0.1ms.
- Toughens the tests to use 20 producers.

Although it became much more rare, it still reboots Macs.
THROTTLE_SLEEP_USEC=10 can pass all tests, but has higher chance of
rebooting, and THROTTLE_SLEEP_USEC=100 does not pass the
sensible_throughput tests.  Increasing the BLOCKSIZE to 16KiB, 32KiB,
1MiB from 4KiB actually has adverse effect on both throughput and
rebooting issues.

This is not an issue on OS X 10.10 iMac and it sounds like the new
MacBook with 10.11 does not reboot either.  MacBook Pros seem to be the
primary victims of this issue.